BoardConfigKernel: Opt-in for bison and flex
For kernel >4.16 bison and flex are required, but there's a catch:
not all of them have a very old dtc compiler and new ones do
actually require a bison version >=3.0 or they will fail to build.

For this reason, let people opt-in for bison and flex: if you
set the env var TARGET_NEEDS_PREBUILT_FLEX_BISON to true, you will
build the components with the prebuilt bison/flex, found in the
Android tree; otherwise, the current system's (your distro's)
bison/flex will be used instead.

Test: Sony Open Devices kernel msm-4.14 now builds ok again.

Change-Id: Id14b2482f4c633f529046b485c6ac16fcd28c45f