PixelPropsUtils: Un-spoof GMS while adding google account [1/2]
GMS uses its DroidGuard service (the same that's used for safetynet)
to get the device name from the build fingerprint, which from our
current certified props, is Pixel XL. The device name after Google
login shows up in the login alert accordingly as Pixel XL, which is
inconvenient especially when you have multiple devices on the same ROM.

By restarting the DroidGuard service process (gms.unstable) as soon as
the user starts adding Google account (i.e when the GMS add account
activity is started and moved to the top of the task stack), we can
skip spoofing to certified props in this instance. When the user
stops/finishes adding account (GMS add account activity is no longer
at the top of the task stack), restart the process again, this time
spoofing to certified props in order to be ready for safetynet.

This ensures while adding a Google account, GMS reads the original
properties of the device and labels the device with its original name,
while in any other situation GMS reads certified properties required
to pass SafetyNet/Play Integrity.

Test:
- Add google account from setup wizard, settings, any other app
- Observe correct device name in gmail login alert
- SafetyNet (basic, CTS), Play Integrity (basic, device) passes.

Change-Id: Ia5463f220a8262f7963fc5b2640a7a5311a73210

* Adapt to current project
* Skip entire gms setProp process by return

Original commit:
https://github.com/ArrowOS/android_frameworks_base/commit/9f490f70b3fb11afa5df6adf7e99a225da563d5a

Change-Id: I428c6753cd6a6cbf820706aea1f2f65932447a19
Signed-off-by: someone5678 <nemui3353@gmail.com>